[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,7 +2,7 @@
 #include <string>    
 using namespace std;
 
-int func(int a, int b){
+long long func(long long a, long long b){
     return (a+b)*(a-b);
 }
 
@@ -10,7 +10,7 @@ int main() {
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
 
-    int a,b;
+    long long a,b;
     cin>>a>>b;
     cout<<func(a,b);
